+++ indico/core/celery/cli.py
@@ -59,6 +59,8 @@ def celery_cmd(args):
                     'INDICO_FLOWER_AUTHORIZE_URL': url_for('oauth.oauth_authorize', _external=True),
                     'INDICO_FLOWER_TOKEN_URL': url_for('oauth.oauth_token', _external=True),
                     'INDICO_FLOWER_USER_URL': url_for('users.authenticated_user', _external=True)}
+        if config.FLOWER_URL:
+            auth_env['INDICO_FLOWER_URL'] = config.FLOWER_URL
         args = ['celery', '-b', config.CELERY_BROKER] + args + auth_args
         env = dict(os.environ, **auth_env)
         os.execvpe('celery', args, env)